+++ Plugins/Creature/Creature.cpp
@@ -119,6 +119,7 @@ Creature::Creature(const Plugin::CreateParams& params)
     REGISTER(LevelDown);
     REGISTER(SetChallengeRating);
     REGISTER(GetAttackBonus);
+    REGISTER(GetHighestLevelOfFeat);
     REGISTER(GetFeatRemainingUses);
     REGISTER(GetFeatTotalUses);
     REGISTER(SetFeatRemainingUses);
@@ -1672,6 +1673,21 @@ ArgumentStack Creature::GetAttackBonus(ArgumentStack&& args)
     }
 
     Services::Events::InsertArgument(stack, retVal);
+    return stack;
+}
+
+ArgumentStack Creature::GetHighestLevelOfFeat(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+    int32_t retval = -1;
+    if (auto *pCreature = creature(args))
+    {
+        const auto feat = Services::Events::ExtractArgument<int32_t>(args);
+          ASSERT_OR_THROW(feat >= Constants::Feat::MIN);
+          ASSERT_OR_THROW(feat <= Constants::Feat::MAX);
+        retval = pCreature->m_pStats->GetHighestLevelOfFeat(feat);
+    }
+    Services::Events::InsertArgument(stack, retval);
     return stack;
 }