+++ CHANGELOG.md
@@ -76,6 +76,7 @@ The following plugins were added:
 - Area: {Get|Set}SunMoonColors()
 - Area: CreateTransition()
 - Creature: GetAttackBonus()
+- Creature: GetHighestLevelOfFeat()
 - Creature: GetFeatRemainingUses()
 - Creature: GetFeatTotalUses()
 - Creature: SetFeatRemainingUses()